requirements.txt
alabaster==0.7.10
Babel==2.4.0
behave==1.2.5
docutils==0.13.1
greenlet==0.4.12
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
msgpack-python==0.4.8
parse==1.8.0
parse-type==0.3.4
peewee==2.9.1
pluggy==0.4.0
psycopg2==2.7.1
py==1.4.33
Pygments==2.2.0
PyHamcrest==1.9.0
pytest==3.0.7
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.5
tox==2.6.0
virtualenv==15.1.0
